requirements.txt
easydict==1.9
jieba==0.39
joblib==0.13.2
Keras==2.3.1
pinyin==0.4.0
tensorflow-gpu==1.14.0
pandas==0.25.0
keras-contrib